feat(client): create useUserFromRedux hook, update user type and getUserProfileAsync reducer, implement it for Profile page
