+++ client/src/pages/Profile/index.tsx
@@ -1,25 +1,19 @@
-import { useEffect } from "react";
 import LogoutButton from "../../components/LogoutButton";
 import SpotifyLink from "../../components/SpotifyLink";
 import { Container, ProfileImage, Text, Title, BoldText } from "./styles";
-import { useSelector, useDispatch } from "react-redux";
-import { RootState, AppDispatch } from "../../state/store";
-import { getUserProfileAsync } from "../../state/spotify/spotifySlice";
+import useUserProfileFromRedux from "./useUserFromRedux";
 
 const Profile = () => {
-  const accessToken = useSelector((state: RootState) => state.auth.accessToken);
-  const dispatch = useDispatch<AppDispatch>();
-  const user = useSelector((state: RootState) => state.spotify.user);
-
-  useEffect(() => {
-    dispatch(getUserProfileAsync(accessToken));
-  }, [accessToken, dispatch]);
+  const { user, loading, error } = useUserProfileFromRedux();
 
   return (
     <Container>
       <Title>User Profile</Title>
-      {user?.display_name ? (
-        /* // TODO: I would support cases when profile could have not all keys from expected */
+      {loading && <Text fontSize="1.1em">Loading profile...</Text>}
+      {error && (
+        <Text fontSize="1.1em">Error: Failed to fetch user profile</Text>
+      )}
+      {user?.display_name && (
         <>
           <ProfileImage src={user?.images[1].url} alt={user?.display_name} />
           <Text fontSize="1em">
@@ -40,9 +34,6 @@ const Profile = () => {
           <SpotifyLink size={32} url={user?.external_urls.spotify} />
           <LogoutButton />
         </>
-      ) : (
-        /* // TODO: if we got an error we always will see loading */
-        <Text fontSize="1.1em">Loading profile...</Text>
       )}
     </Container>
   );
