+++ client/src/types.ts
@@ -14,6 +14,10 @@ export interface UserProfile {
   product: string;
   type: string;
   uri: string;
+  error: {
+    status: number;
+    message: string;
+  };
 }
 
 interface Image {
